Added global box-sizing override
This is probably the main thing that was catching you up. By default
everything is `box-sizing:content-box` which means when you set a width
on an element, it's being applied to the content-box of the element,
which is the width before any padding/border/margin is added to the
size, and generally is not the desired result. I add this little snippet
to everything I do.